Merged [13260]: Cache the run loop used for the gaim thread, and use it when adding sources and timeouts.  This yields a minor efficiency boost (it's an operation we do quite a  bit, and the result of CFGetCurrentRunLoop() should always be the same).
Additionally, it fixes a situation in which it's not the same.  gaim_account_set
_int() and cousins schedule an account.xml save... and we call those functions o
n the main run loop, which previously led to a scheduled save on the main run lo
op, which could conflict with g_mem activity on the gaim thread and cause crashe
s in rare situations.  Such scheduled saves are no shuffled to the gaim thread.