Outline and the most important functionality (account setup) of the setup assistant.
Feel free to write good text for the three existing panes and to add some sort of basic preferences configuration pane for it :)

There's a #define in AIAdium.m {{{ALWAYS_RUN_SETUP_WIZARD}}} which can be changed to TRUE to see it with an existing configuration.  Otherwise, it runs when no accounts are configured at startup.

The !AdiumyButler image is by Wes Underwood, of course.

Refs #9.