60% or so of an improved log viewer with iTunes-browser style tables at the top.  I'm committing it now mostly to get feedback on its direction and the improvements thus far.
I rejected the Mail.app-search-style top type (from, to, content, etc.) bar because it felt too restricting, but I'm open to suggestions.

Note that the (Date) column is not implemented.  Rather than listing every possible date, I picture it having "Today", "Yesterday", "This Week", "This Month", and so on.

In the process, much of the old log viewer moved to AIAbstractLogViewerWindowController, which is really only abstract in the sense that it is a common base class for the two log viewers (10.3 and 10.4) we now support, appropriate since the bottom half of the new one looks like and works like the old one.

Along with this commit we also see a huge speed-up in loading of logs in the old-fashioned, non-searchkit way (which is still used, in a modified form, for an initial 'search' of no criteria so is relevant on 10.4).

Refs #38.