Modified [15593] to still implement custom view settingFrom a post to adium-devl: [15593] breaks custom views in toolbars when they are instantiated via the AIToolbarUtilities classes.  I completely forgot about this changeset and just spent 2 hours being mystified as to why the search field in a toolbar was behaving as it were a different view than the one I instantiated and stuck in the view; the reason was that NSView doesn't conform to NSCopying.  Our custom copying behavior sets the view on the new toolbar item; the non-custom copying behavior instantiates a new view of the same class and attempts to make it match.
Technically, using the default copying behavior is Better... but then we have to do all sorts of checking elsewhere if we stick a custom view into a toolbar item and expect to be able to access that view from elsewhere.  I've committed a compromise which I think keeps your bug fixed but also makes custom views which can't be copied do right.