Read and write dictionaries via {{{NSPropertyListSerialization}}}, using the binary plist file ({{{NSPropertyListBinaryFormat_v1_0}}}) for writing.
I did some Sharking on our preference loading and determined that the majority of the time was spent actually reading the bytes of the plist file off the disk. The binary format uses about 50% of the disk space... and so has 50% less bytes to read and to write.  I believe this cost-benefit analysis -- HD access is more expensive than CPU access -- is why Apple switched to using the binary format for NSUserDefaults in 10.4.

Previous code reads the binary plists just fine (so preferences remain readable by Adium 0.8x).