Fixed reconnecting after disconnecting in many situations for Gaim accounts, particularly after being unexpectedly disconnected.
The problem: when you go offline, Gaim keeps your old status as 'active'.  Adium does something similar.  However, if you then try to just set the status again, without changing anything (which happens when you're being reconnected by Adium - it keeps the status the same as it was before), the Gaim core says "already active, don't need to change anything" and goes about its merry way... without connecting.

The fix: If we set a status which is an online status, and the applicable account is neither online nor connecting after we do so, tell the account to connect.

This was for some reason a particular problem for MSN accounts... but it should have been a problem for all libgaim ones. Fixes #4304 (of which at least 14 tickets have been marked a duplicate), Refs #3147 (the original ticket is about a smarter reconnect process, but this fixes some issues detailed in its comments).