Removed the System Version Check launch utility and build target, and set our minimum required system version to 10.3.9 in the Apple-approved way.  This is ''broken'' in 10.3, which means that 10.3.0 through 10.3.8 users will be greeted with a cryptic error message rather than an appropriate one.  It works in 10.2, which means that 10.2 users will be told that this version of Adium does not support their operating system.
Based on reports from Gareth Potter, System Version Check was responsible for failure to launch under all our Asian localizations.  I'm still not sure why, but assuming that was the real cause, this fixes #4843 -- I'd like verification of that when 1.0b9 is released.

This should fix sporadic reports of "multiple adiums are running, one doesn't have a menu bar" or the like.. again, strange, but seemingly linked to the crash reporter relaunching Adium without quitting the previous Adium.real.

This also fixes a reported problem launching in the 10.5 preview.