This should fix [http://www.visualdistortion.org/crash/view.jsp?crash=200795 this crash], which occurred in {{{-[AIKeychain setInternetPassword:forServer:securityDomain:account:path:port:protocol:authenticationType:keychainItem:error:]}}} via {{{-[AdiumPasswords setPassword:forAccount:]}}}.
We use a local autorelease pool in the keychain method, in order that certain objects that are in scope there get released quickly. Unfortunately, this pool was scooping up the {{{NSError}}} object as well, since it's autoreleased. We now retain the {{{NSError}}} and release it after the pool has been popped.