Fixed the bug that caused b10 to be pulled.
How did this happen? (a bit of analysis)
When I removed the pseudo-KVC methods that were causing the runtime to complain (see the bottom of AIXMLElement.m), I compiled, ran some code I knew depended on AIXMLElement (the in-progress message history), and figured it was working, so I committed.
The problem was that setValue:forKey and valueForKey: are defined, as far as the compiler is concerned, so their use in AIHTMLDecoder was masked. Looking at the output of Adium in gdb, I quickly realized the error, wrote some real accessors, tested them, and now I'm committing.
Sorry about this, folks.