This is libgaim.framework at r 315, which is Gaim to revision 17191 except proxy.c which is at 17164
Worth noting: gaim revision 17180 (kingant: "Don't allow dns lookups of empty strings") did not fix the crash in strncopy() we get after certain MSN profile lookups - it is still present in this build of libgaim.

That said, with proxy.c at pre-17165, we can connect to Yahoo and Jabber again.  I think we're good to go for 1.0bwhatvernumberwe'reonnowdangitiwishwecouldmakeeverythingrightandget1.0outthedoor.