Okay, we're getting closer.  Gaim can in some situations request to add an input on a source when one already exists -- for example, a write input on an fd for which we already have a write input.  I suspect this is an underlying bug.  It looks like one place it can happen is at the end of the UPNP IP detection process, which might explain why it is seemingly randomly distributed who sees this bug (and why it happens consistently and I don't see it locally, since Apple Airport Express doesn't support UPNP).
In any case, the proper behavior in that situation is to remove the old input and replace it with the new one.  We must be sure to remove its tag from the {{{sourceInfoDict}}} which tracks such things, because Gaim may at some point later request to remove it, which is where we crash because we pull out freed info and go wild.

Refs #5510