Sigh.  Alas, poor joscar! I knew him, Horatio.
As discussed on adium-devl, joscar was a wonderful and terrible experiment.  Wonderful in its promise; terrible that it was doomed to fail through no fault of its own.  The Java-Cocoa Bridge, the foundation of our communication with the Java libraries, is deprecated -- promised to fail in 10.6, failing in current 10.5 builds, and buggy in both 10.4 and 10.3.  It is no fit groundwork for Adium no matter how beautiful the structure it supported.

Fortunately, while we've been working on joscar for AIM for a year, libgaim hasn't been sleeping.  File transfer and Direct IM work quite nicely there, and recently Mark Doliner implemented joscar's rate limiting management algorithm there, as well.  It's a good place to be.