Give AIAccount subclasses an opportunity not to send a status update as soon as the account connects. While this is reasonable default behavior in general, it is not only duplicative for libgaim, since we specify the status information during the connect process, but also detrimental, since the protocol may wish to wait to send such information until a later time.  Case in point: in Jabber, we don't want to send presence information until the roster has been requested, which does not occur until server capabilities discovery is complete.
Also, properly filter and set the status message for an account which is not yet in the online state.  This fixes a minor bug for libgaim accounts in which the status message would not be set until an instant after connection instead of during the connection process if the protocol supports that... a minor bug which became major and therefore a higher priority with the fix above.