I'm not clear how metacontacts signing on while View->Show Groups is not active has *ever* worked...  1. A metacontact which isn't on the list at present because none of its contacts are on the lists of connected accounts has a group of nil.  2. When a contact signs on and is assigned a remote group, it notifies its metacontact via {{{-[AIMetaContact remoteGroupingOfContainedObject:changedTo:}}}.  3. If the metacontact isn't in a group, it then calls {{{-[self restoreGrouping]}}}...  4. which previously had no action if we aren't using contact list groups.
At (4), we now put the metacontact in the contact list root if we're not supposed to be using groups.  We also check so that in (3) to ensure that ''if'' we're not using contact list groups ''and'' the metacontact isn't already in the root we can call restoreGrouping.

jmpp reports that the issue only occurred for him while not showing list groups; I bet the same is true for the other reports of an identical issue in the ticket. Based on his testing, fixes #5901.