Merged [18926]: Some exceptions at the top-level are not meant to ever be considered, it seems. NSAccessibilityExpcetion exceptions, for example, are thrown regularly for controls which don't have full accessibility support... but they are silently dropped normally.  Our overzealous exception handling was (1) logging them to console and (2) going to pains to get the stack trace out of them, a slow process.
For exceptions we choose to ignore, we now fully ignore them.  Additionally, if the exception has a name or reason known to be worth ignoring, we don't waste effort generating the stack trace.

This fixes problems with Adium + VoiceOver causing slowdowns. Fixes #6904.

I'm still not sure what's up with the "-[NSArray addObject:] attempt to add nil" crashes which are also somehow tied into VoiceOver.