Libgaim.framework to r 401 plus some local changes which I'll commit to libgaim if they work out well.
Among other fixes, this has my patch in #6378 committed upstream such that children created via fork() don't have the signal handlers installed... so a crash or signal in a child shouldn't attempt to launch the crash reporter and end up a permanent drain on resources.  The DNS resolve in question still, fails, though, seemingly as a result of a request in underlying code to malloc a ridiculous number of bytes.  The console log in #6276 (which address the same issue as #6378 but with different data) says:
{{{
*** malloc_zone_calloc[379]: arguments too large: 1,4293749504
}}}
and malloc_zone_calloc() is where the child process is indeed crashing.
That problem should be addressed separately, I believe. Closes #6276; closes #6378.

This also commit moves adium-1.0 to the new adiumGaimEventLoop implementation I already committed to trunk in [18985]; it works well for me, and it's easier to work with overall.  It also adds an implementation of the new event loop ui op input_get_error() which based on discussion Graham and I have had should let us fix the peculiarities in which GTK-Gaim and Adium socket read/write behaviors differ, causing things like AIM file transfer failure when the same transfer works in Gaim.