One of the longest times to find a one-line fix in my experience.
An NSTabViewItem retains its identifier. AIMessageTabViewItem, as of PSMTabBarControl, sets itself as identifier.  This should be refactored at some point; for now, we'll set the identifier to nil after removing it from its NSTabView in AIMessageWindowController, allowing it to ultimately deallocate (taking with it the rest of its retained hierarchy).

This fixes the leak [http://adiumx.com/pipermail/adium-devl_adiumx.com/2007-April/002656.html I described earlier on adium-devl] and in [19459].