Added DiskImageUtilities and AuthorizedTaskManager from Google's MacFUSE repository; they are BSD licensed.  DiskImageUtilities is used in AIAdium to check if we are running from our read-only installation disk image and, if so, to prompt the user as to whether s/he would like to copy to Applications instead.  If the user agrees, Adium is copied to Applications and relaunched from there.  If the user declines, no further prompting is done.
AuthorizedTaskManager is used to authenticate the user for writing to Applications if necessary.

The copy process automatically overwrites any existing copy of Adium.app in /Applications; the dialogue warns the user of this.

Much thanks to Avi Drissman and Google for providing this code which is originally from [http://toolbar.google.com/gmail-helper/notifier_mac.html Google Notifier] and was relicensed as BSD within the MacFUSE repository for our use :)