Fixed display issue for the contact list outline view when it is initially opened with a non-empty list.
When the list reloads, it caches the row heights for items.  If it reloads before groupCell and contentCell are created and assigned, we get a 0 height, which we were returning as 16. That just masks the problem and so turned this into an odd "incorrect cell height" display issue (16 instead of whatever the height is supposed to be).

Oddly, {{{	[self noteHeightOfRowsWithIndexesChanged:[NSIndexSet indexSetWithIndexesInRange:NSMakeRange(0, [self numberOfRows])]];}}} is called in AIMultiCellOutlineView's setContentCell and setGroupCell, but NSOutlineView seems to ignore this and not update its cache.