Optimize how we deal with Message Styles. Previously we would create one per chat, including loading the HTML templates and the icon mask image from disk each time. Also, we would preload the list of available message style on a background thread at launch time. This required the use of @synchronized to protect said list.
This patch caches the path to the current message style in our prefs system, so it can avoid scanning the entire message style folder, and eliminate the need for preloading the list. It also creates one style object per style, rather than per chat, which should save a good chunk of ram. On top of all this, there's a leak fix, since the style object didn't release one of its ivars.

I am seeing an intermittent issue when changing styles in prefs, so further testing is definitely welcome.