Burninated 20 headers that had been snuck into the !CopyFiles phase for the Spotlight importer.
Watch for this when you add headers to the project, guys. It's a waste of valuable time for Xcode to copy a bunch of headers, only for them to be bulldozed when our script carpet-bombs every header in the application. (That script is intended for headers that come with frameworks; there shouldn't be headers in the application for it to remove.)