Fixed regression that I introduced in r20428, wherein you couldn't use multiple sounds loaded into the cache.
This was caused by my attempt to reuse a single audio context amongst all the cached sounds; !QuickTime apparently rejects such frugality. We now create an audio context, set it in the movie, and then forget about it; we don't try to reuse contexts.

As a side effect of this fix, the code is much simpler.