Stopping the animation doesn't lead to `animationDidEnd:` being called, so our `_reallyCloseToolTip` method was never called, so the current tooltip was not released if another one was spawned while the current one was fading out. This caused tooltips to be leaked in `_closeTooltip`, when we replaced the existing animation id with a new one, boldly assuming that one wasn't there already.
To fix this, we must implement `animationDidStop:` as well as `animationDidEnd:`. Currently, the implementations are equal. I also added an assertion to `_closeTooltip` so that conditions like this one are caught in the future.

I believe that this fixes #7585. Committed to both [source:trunk trunk] and the [source:branches/adium-1.1 1.1 branch].