So the message window controller implicitly retains the window, 	and the window retains its content view, 	and the content view retains its subviews, including the split view, 	which retains its subviews, including the tab view of chats, 	which retains its tab view items, 	which retained the message window controller.
It doesn't look like this was an actual leak, which I think is thanks to windowWillClose: closing all the chats of the window, but circular retains are bad anyway.

This change partially reverts r7795 (it reverts the intended change, but does not revert another change that looks like it got swept up into that).