Okay, I'm clearly overtired. In [20854], I committed a patch from Reid Barton to fix #5571.  In [21290], I committed a patch from Pheredhel to fix #6610.  #6610, however, is a duplicate of #5571...
Caching the preference value, as Pheredhel does in #6610's patch, is a good idea.  However, that patch is actually broken in that it disables the overlay entirely rather than preventing animation in the case that animation should not happen.  This was being handled properly by the #5571 patch...

The result, now, is a combination of the two. Apologies to all for the noisy commits. Refs #5571. Refs #6610.