Merged [21399] and [21400]: Disabled our hackery which was intended to let windows ignore expose, useful only when ordering below other windows and then revealing the desktop. This was causing intermittent problems with Spaces in 10.5 and simply isn't worth it IMO. I've left the code in place but commented out in case someone thinks its worth working towards a fix which can maintain the old feature.
This fixes all the cases which I could reproduce before with Spaces problems, with the exception that windows which are set to order above other windows do not switch spaces when dragged-and-held at the screen edge. This seems to be either a Spaces bug or intentional behavior. I've filed it as radr://5565888

Fixes #7862.