Added a patch to libpurple, application of the patch, and a built libpurple im.pidgin.pidgin @ 5727e9151416612d63670c49ed8982d3bb1704b4 which has the patch applied.  This adds a libpurple preference (code only, not in the UI) which asks xmpp to use its own PLAIN and DIGEST-MD5 implementations before trying cyrus-sasl.  We only set this to TRUE if the client is running on 10.4, where PLAIN cyrus-sasl fails authentication because the username is mysteriously doubled before Base64 encoding is performed.
The result is that on 10.5+, we use cyrus-sasl exclusively, which then figures out for us the best authentication method and uses it. That can include much more complex methods such as Kerberos5/GSSAPI, needed to connect to iChat Server 10.5 among others, or CRAM-MD5.

On 10.4, we use libpurple's DIGEST-MD5 implementation if the server supports that.  Failing that, we use libpurple's PLAIN if the server supports that. Failing that, we pass the attempt on to cyrus-sasl.  This -might- work for the other mechanisms, but we have no testing on that so far.

Fixes #8113 (I hope!).