If we have an incoming connection from an unknown contact, create a AWEzvContact for it and add it to the contacts dictionary and then proceed as normal rather than dropping the connection and logging an error.  We may end up with a contact with an unknown status message or missing buddy icon, but that's better than not receiving a message or file transfer at all.
I haven't been able to reproduce the situation in which this could happen, but #8116 shows logs of it in the wild.  Fixes #8116.

Also, I removed the confusing back-and-forth switching of `contactIdentifier` between a dotted-quad (IP address) and a hyphened-quad (contact identifier key for the contacts dictionary).  Two variables, `ipAddr` and `contactIdentifier`, are now used to make the code clearer.