Merged the contact-icon-management-rewrite branch to trunk.
This expands AIUserIcons to become to manager of contact icons.  Objects conforming to the AIUserIconSource protocol register with it and inform AIUserIcons of AIListObject's icons on demand; AIUserIcons manages a cache of this information.  We currently a Manual icon source, Serverside icon source, Cached icon source, and Address Book icon source.  The order in which they are used is determined by their priority; only the Address Book has a configurable priority, which is based upon the "even if the contact already has an icon" checkbox in the advanced Address Book preferences.

Previously, a single AIListContact could have as many as 3 different NSImage objects associated with it just to have a single user icon because they were all kept around and ordered by priority.  Furthermore, multiple AIListContacts (e.g. 'tekjew' on two different AIM accounts) rarely shared NSImage files even if they had the same actual image.  Finally, all possible sources of icons were previously queried and used even though all but one would only be used if the highest-priority one were removed.  All these things are fixed in the new implementation.

The 'cached' user icon source is actually not going to be used for most of Adium now.  Accounts can now specify that they manage their own cache of serverside icons, and we now use the libpurple cache (which was being written out and tracked even when we weren't using it) and supply its icons as needed.  This means that cached icons are treated as if they are serverside icons - which is fine behavior and more straightforward.  Opening a chat window will now always need to write out the icon once to be able to pass it to the WebView  (previously it could make use of the path to the cached icon if there was one), but that's a very minimal one-time price to pay and may resolve some very rare problems in which the icon would be out of date in the WebView, anyways.