Merged r22649 from trunk to 1.2 branch:
Halfway reverted r22592, which broke sign-in for people using @sbcglobal.net addresses with Yahoo!. We now delete the suffix only if it starts with ?@yahoo.?, which should preserve the spirit of r22592.

Failure cases are any suffix that is a valid non-Yahoo! domain name that starts with ?yahoo.?; for example, ?@yahoo.example.com?. Hopefully, these are rare enough (if they exist at all) that this won't be a problem.