Fixed the build under Xcode 2.4.x. Fixes #9320.
Apparently, this variable is empty under Xcode 2.4.x. This causes the whole definition to be empty, resulting in a command line containing ?-F -filelist path/to/file.FileList?, which results in the ?missing argument to: -filelist option? error message. (I guess it gives the pathname to the first argument that needed one, which would be -F.)

The fix is simply to add a dummy pathname to satisfy the constraint that there must be at least one.