Added support for a text-color checkbox to the text color preview view. If such a checkbox is present, and it's '''un'''checked, the view will use black instead of the text color from its color well.
Thus, the three text color preview views on the Window tab of the List Theme sheet now use the Online text color on the Status tab only if the Online colors are turned on. If the box is unchecked, the views on the Window tab use black, just as the Contact List does.