Detect changes to aliases directly, make use of the server alias over the local alias since we know our own local alias and keep track of it separately, and generally be far less ugly in how we look for alias updates. Refs #6828, as coupled with the next commit it will be fixed.
Also, note when new contacts are added directly rather than waiting on the generic blist update, and register our signals and accounts soon enough to be able to make use of the blood of useful information as the blist is loaded.  The result is a far snappier perceived sign-on experience, since we're not scuffing our boots in the dirty waiting for libpurple to accidentally repeat what we should have heard the first time 'round.  It's amazing what using your API instead of fighting it can accomplish...