Merged [23115]: We'll be running message through an HTML decoder. If we unescape all entities, we end up treating any typed HTML code as a real HTML code, because for example `&lt;/html&gt;` becomes </html> before the decoder even sees it. That's bad.
We apparently need to unescape ampersand immediately (I didn't retest that claim, made in #6850).  We'll just do that rather than a full unescaping.  Note that we can't use stringByUnescapingFromXMLWithEntities: with a dictionary specifying just the ampersand, the documentation for `CFXMLCreateStringByUnescapingEntities()`, which that method wraps, states "The standard XML entities are always replaced".

Fixes #9591