This is 95% Colin's work; I put some finishing touches on it.
Renamed the 'status object' system to the 'property' system.  Instead of having "status objects" for "keys" associated with contacts, accounts, and chats, which often had little to do with 'status', we now have values for "properties."  This matches the wordage used in Cocoa's KVC system and generally feels better.

This should be merged to the unit-tests branch sooner than later to prevent merging difficulties down the line; I wanted it in before we branched for 1.4svn or for GSoC 2008.